\name{facopy.annot-package}
\alias{facopy.annot-package}
\docType{package}
\title{
Companion annotation package for facopy
}
\description{
Provides facopy with genome annotation on chromosome arms, genomic features and copy number alterations.
}
\details{
\tabular{ll}{
Package: \tab facopy.annot\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-08-27\cr
License: \tab GPL-3\cr
}
}
\author{
David Mosen-Ansorena
}

\keyword{ package }

