\name{facopy_biocarta}
\alias{facopy_biocarta}
\docType{data}
\title{
Biocarta Pathways with symbol identifiers
}
\description{
Modification of the \code{biocarta} object in \code{graphite} package, in order to list gene symbols instead of the native identifiers.
}
\source{
graphite R package.
}
\references{
Sales, G., Calura, E., Cavalieri, D. & Romualdi, C. graphite - a Bioconductor package to convert pathway topology to gene network. BMC Bioinformatics 13, 20 (2012). 
}
\keyword{datasets}
