\name{facopy_msigdb}
\alias{facopy_msigdb}
\docType{data}
\title{
facopy MSigDB Data
}
\description{
Contains gene sets, classified into collections.
}
\source{
MSigDB
}
\references{
Liberzon, A. et al. Molecular signatures database (MSigDB) 3.0. Bioinformatics 27, 1739-40 (2011).
}
\keyword{datasets}
