\name{hg18_armLimits}
\alias{hg18_armLimits}
\docType{data}
\title{
hg18_armLimits
}
\description{
Chromosome arm upper limits (in base pairs) for the hg18 genome build.
}
\usage{data(hg18_armLimits)}
\format{
  A data frame with 48 observations on the following 2 variables.
  \describe{
    \item{\code{chr_q_arm}}{A factor with levels \code{1p} \code{1q} \code{2p} \code{2q} \code{3p} \code{3q} \code{4p} \code{4q} \code{5p} \code{5q} \code{6p} \code{6q} \code{7p} \code{7q} \code{8p} \code{8q} \code{9p} \code{9q} \code{10p} \code{10q} \code{11p} \code{11q} \code{12p} \code{12q} \code{13p} \code{13q} \code{14p} \code{14q} \code{15p} \code{15q} \code{16p} \code{16q} \code{17p} \code{17q} \code{18p} \code{18q} \code{19p} \code{19q} \code{20p} \code{20q} \code{21p} \code{21q} \code{22p} \code{22q} \code{Xp} \code{Xq} \code{Yp} \code{Yq}}
    \item{\code{limit}}{A numeric vector}
  }
}
\examples{
data(hg18_armLimits)
}
\keyword{datasets}
