\name{mm8_armLimits}
\alias{mm8_armLimits}
\docType{data}
\title{
mm8_armLimits
}
\description{
Chromosome arm upper limits (in base pairs) for the mm8 genome build.
}
\usage{data(mm8_armLimits)}
\format{
  A data frame with 21 observations on the following 2 variables.
  \describe{
    \item{\code{chr_q_arm}}{A factor with levels \code{1q} \code{2q} \code{3q} \code{4q} \code{5q} \code{6q} \code{7q} \code{8q} \code{9q} \code{10q} \code{11q} \code{12q} \code{13q} \code{14q} \code{15q} \code{16q} \code{17q} \code{18q} \code{19q} \code{Xq} \code{Yq}}
    \item{\code{limit}}{A numeric vector}
  }
}
\examples{
data(mm8_armLimits)
}
\keyword{datasets}
